requirements.txt
# Core Framework
fastapi==0.115.14
uvicorn==0.35.0
starlette==0.46.2
pydantic==2.11.7
pydantic_core==2.33.2

# Database
SQLAlchemy==2.0.41
alembic==1.16.5
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.21.0
greenlet==3.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==45.0.6
python-multipart==0.0.20

# Validation & Utils
email-validator==2.2.0
python-decouple==3.8
python-dotenv==1.1.1
fastapi-pagination==0.14.0

# Testing
pytest==8.4.1
httpx==0.28.1

# Dev Tools
black==25.1.0
rich==14.0.0
